ftrace_reader: Add dynamic parser (part 1)
To dynamically parse ftrace events into protos we need to merge some
build time information (proto field ids, proto field types) with
some runtime information (ftrace field types). This CL:
- Refactors ProtoTranslationTable to accept partially filled in Event
  structs rather than generating them from scratch.
- Adds event_info.cc, which holds the (eventually auto generated)
  build time info.
- Also happens to fix the issue with available_events (since we
  only need to look for the events we have static information
  for).

Bug: 69662589,70328826

+++ include/perfetto/ftrace_reader/ftrace_to_proto.h
@@ -30,8 +30,8 @@ namespace perfetto {
 struct FtraceEvent {
   struct Field {
     std::string type_and_name;
-    size_t offset;
-    size_t size;
+    uint16_t offset;
+    uint16_t size;
     bool is_signed;
 
     bool operator==(const Field& other) const {
